Add argument parsing from CLI
For better backwards compatibility arguments are named consistently with
their corresponding config options.

Config is read in over the top of any arguments that are set to their
defaults so that anything specified on the command line takes
precedence.

RESTRICT has been dropped due to redundancy.

`--header` has been renamed to `--header-preset` which loads one of the
    built-in user agent/header sets.

`--header` accepts a key-value pair delimited by an equals (`=`) sign.
It can be specified multiple times to allow setting an arbitrary number
of headers. The config file format accepts a python-style list e.g.
`header = [a=b, c=d, ...]`